+++ src/sdk/url-strategy.ts
@@ -97,4 +97,42 @@ function urlStrategyRetries<T>(
   }
 }
 
-export { urlStrategyRetries, UrlStrategy, BaseUrlsMap }
+interface BaseUrlsIterator extends Iterator<BaseUrlsMap> {
+  reset: () => void;
+}
+
+function getPreferredUrls(endpoints: Partial<Record<UrlStrategy, BaseUrlsMap>>): BaseUrlsMap[] {
+  const preferredUrls = getEndpointPreference()
+
+  if (!Array.isArray(preferredUrls)) {
+    return [preferredUrls]
+  } else {
+    const res = preferredUrls
+      .map(strategy => endpoints[strategy] || null)
+      .filter((i): i is BaseUrlsMap => !!i)
+
+    return res
+  }
+}
+
+function getBaseUrlsIterator(endpoints: Partial<Record<UrlStrategy, BaseUrlsMap>> = endpointMap): BaseUrlsIterator {
+  const _urls = getPreferredUrls(endpoints)
+
+  let _counter = 0
+
+  return {
+    next: () => {
+      if (_counter < _urls.length) {
+        return { value: _urls[_counter++], done: false }
+      } else {
+        return { value: undefined, done: true }
+      }
+    },
+    reset: () => {
+      _counter = 0
+    }
+  }
+}
+
+
+export { urlStrategyRetries, getBaseUrlsIterator, BaseUrlsIterator, UrlStrategy, BaseUrlsMap }