+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+### Version 5.3.0 (25th March 2021)
+#### Added
+- Added `getWebUUID` method to access SDK generated ID `web_uuid`.
+- Added `getAttribution` method to access user's current attribution information.
+
+#### Fixed
+- Fixed issue with URL strategy retrying to send requests after SDK was disabled.
+
+---
+
 ### Version 5.2.1 (16th September 2021)
 #### Fixed
 - Fixed top-level Typescript declarations.